requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
cryptography
python-dotenv
pydantic
apify-client
apscheduler
alembic
tzdata
httpx
